llm-service/requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
grpcio-health-checking==1.68.1
protobuf==5.29.2
requests==2.32.3
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-grpc==0.50b0
opentelemetry-instrumentation-requests==0.50b0
